requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
Flask-Migrate
beautifulsoup4 
schedule
requests
Flask-APScheduler==1.13.1
PyJWT